+++ HISTORY.rst
@@ -5,5 +5,8 @@ History
 0.1.0 (2018-06-01)
 ------------------
 
+* Optimise memory usage (:pr:`9`)
+* Support ignoring initial flags (:pr:`9`)
+* Support flagging on Polarised Intensity (:pr:`9`)
 * Added YAML configuration (:pr:`4`)
 * Added a progress bar (:pr:`3`)